server: refactored TLS utilities to separate package
