+++ cli/command_server_tls.go
@@ -2,22 +2,20 @@ package cli
 
 import (
 	"context"
-	"crypto/rand"
 	"crypto/rsa"
 	"crypto/sha256"
 	"crypto/tls"
 	"crypto/x509"
-	"crypto/x509/pkix"
 	"encoding/hex"
-	"encoding/pem"
 	"fmt"
-	"math/big"
 	"net"
 	"net/http"
 	"os"
 	"time"
 
 	"github.com/pkg/errors"
+
+	"github.com/kopia/kopia/internal/tlsutil"
 )
 
 const oneDay = 24 * time.Hour
@@ -32,88 +30,11 @@ var (
 )
 
 func generateServerCertificate(ctx context.Context) (*x509.Certificate, *rsa.PrivateKey, error) {
-	log(ctx).Infof("generating new TLS certificate")
-
-	priv, err := rsa.GenerateKey(rand.Reader, *serverStartTLSGenerateRSAKeySize)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "unable to generate RSA key")
-	}
-
-	notBefore := time.Now()
-	notAfter := notBefore.Add(time.Duration(*serverStartTLSGenerateCertValidDays) * oneDay)
-
-	serialNumber, err := rand.Int(rand.Reader, new(big.Int).Lsh(big.NewInt(1), 128))
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "unable to generate serial number")
-	}
-
-	template := x509.Certificate{
-		SerialNumber: serialNumber,
-		Subject: pkix.Name{
-			Organization: []string{"Kopia"},
-		},
-		NotBefore: notBefore,
-		NotAfter:  notAfter,
-
-		KeyUsage:              x509.KeyUsageKeyEncipherment | x509.KeyUsageDigitalSignature,
-		ExtKeyUsage:           []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth},
-		BasicConstraintsValid: true,
-	}
-
-	for _, n := range *serverStartTLSGenerateCertNames {
-		if ip := net.ParseIP(n); ip != nil {
-			log(ctx).Infof("adding alternative IP to certificate: %v", ip)
-			template.IPAddresses = append(template.IPAddresses, ip)
-		} else {
-			log(ctx).Infof("adding alternative DNS name to certificate: %v", n)
-			template.DNSNames = append(template.DNSNames, n)
-		}
-	}
-
-	derBytes, err := x509.CreateCertificate(rand.Reader, &template, &template, priv.Public(), priv)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "failed to create certificate")
-	}
-
-	cert, err := x509.ParseCertificate(derBytes)
-	if err != nil {
-		return nil, nil, errors.Wrap(err, "failed to parse certificate")
-	}
-
-	return cert, priv, nil
-}
-
-func writePrivateKeyToFile(fname string, priv *rsa.PrivateKey) error {
-	f, err := os.Create(fname)
-	if err != nil {
-		return err
-	}
-	defer f.Close() //nolint:errcheck,gosec
-
-	privBytes, err := x509.MarshalPKCS8PrivateKey(priv)
-	if err != nil {
-		return errors.Wrap(err, "Unable to marshal private key")
-	}
-
-	if err := pem.Encode(f, &pem.Block{Type: "PRIVATE KEY", Bytes: privBytes}); err != nil {
-		return errors.Wrap(err, "Failed to write data to")
-	}
-
-	return nil
-}
-
-func writeCertificateToFile(fname string, cert *x509.Certificate) error {
-	f, err := os.Create(fname)
-	if err != nil {
-		return err
-	}
-	defer f.Close() //nolint:errcheck,gosec
-
-	if err := pem.Encode(f, &pem.Block{Type: "CERTIFICATE", Bytes: cert.Raw}); err != nil {
-		return errors.Wrap(err, "Failed to write data")
-	}
-
-	return nil
+	return tlsutil.GenerateServerCertificate(
+		ctx,
+		*serverStartTLSGenerateRSAKeySize,
+		time.Duration(*serverStartTLSGenerateCertValidDays)*oneDay,
+		*serverStartTLSGenerateCertNames)
 }
 
 func startServerWithOptionalTLS(ctx context.Context, httpServer *http.Server) error {
@@ -151,13 +72,13 @@ func maybeGenerateTLS(ctx context.Context) error {
 
 	log(ctx).Infof("writing TLS certificate to %v", *serverStartTLSCertFile)
 
-	if err := writeCertificateToFile(*serverStartTLSCertFile, cert); err != nil {
+	if err := tlsutil.WriteCertificateToFile(*serverStartTLSCertFile, cert); err != nil {
 		return errors.Wrap(err, "unable to write private key")
 	}
 
 	log(ctx).Infof("writing TLS private key to %v", *serverStartTLSKeyFile)
 
-	if err := writePrivateKeyToFile(*serverStartTLSKeyFile, key); err != nil {
+	if err := tlsutil.WritePrivateKeyToFile(*serverStartTLSKeyFile, key); err != nil {
 		return errors.Wrap(err, "unable to write private key")
 	}
